+++ src/lib.rs
@@ -73,6 +73,7 @@
 //! - `dataloader`: Support [DataLoader](dataloader/struct.DataLoader.html).
 //! - `decimal`: Integrate with the [`rust_decimal` crate](https://crates.io/crates/rust_decimal).
 //! - `cbor`: Support for [serde_cbor](https://crates.io/crates/serde_cbor).
+//! - `smol_str`: Integrate with the [`smol_str` crate](https://crates.io/crates/smol_str).
 //!
 //! ## Integrations
 //!