+++ tests/validators.rs
@@ -341,6 +341,20 @@ pub async fn test_custom_validator() {
         async fn input(&self, input: MyInput) -> i32 {
             input.n
         }
+
+        async fn value2(
+            &self,
+            #[graphql(validator(list, custom = "MyValidator::new(100)"))] values: Vec<i32>,
+        ) -> i32 {
+            values.into_iter().sum()
+        }
+
+        async fn value3(
+            &self,
+            #[graphql(validator(list, custom = "MyValidator::new(100)"))] values: Option<Vec<i32>>,
+        ) -> i32 {
+            values.into_iter().flatten().sum()
+        }
     }
 
     struct Subscription;
@@ -443,6 +457,54 @@ pub async fn test_custom_validator() {
             path: vec![PathSegment::Field("value".to_string())],
             extensions: None
         }]
+    );
+
+    assert_eq!(
+        schema
+            .execute("{ value2(values: [77, 88] ) }")
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "[Int!]": expect 100, actual 77"#.to_string(),
+            source: None,
+            locations: vec![Pos {
+                line: 1,
+                column: 18
+            }],
+            path: vec![PathSegment::Field("value2".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute("{ value3(values: [77, 88] ) }")
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "[Int!]": expect 100, actual 77"#.to_string(),
+            source: None,
+            locations: vec![Pos {
+                line: 1,
+                column: 18
+            }],
+            path: vec![PathSegment::Field("value3".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute("{ value3(values: null ) }")
+            .await
+            .into_result()
+            .unwrap()
+            .data,
+        value!({
+            "value3": 0
+        })
     );
 }
 
@@ -603,5 +665,127 @@ pub async fn test_validate_wrapper_types() {
             path: vec![PathSegment::Field("f".to_string())],
             extensions: None
         }]
+    );
+}
+
+#[tokio::test]
+pub async fn test_list_both_max_items_and_max_length() {
+    struct Query;
+
+    #[Object]
+    impl Query {
+        async fn value(
+            &self,
+            #[graphql(validator(list, max_length = 3, max_items = 2))] values: Vec<String>,
+        ) -> String {
+            values.into_iter().collect()
+        }
+
+        async fn value2(
+            &self,
+            #[graphql(validator(list, max_length = 3, max_items = 2))] values: Option<Vec<String>>,
+        ) -> String {
+            values.into_iter().flatten().collect()
+        }
+    }
+
+    let schema = Schema::new(Query, EmptyMutation, EmptySubscription);
+    assert_eq!(
+        schema
+            .execute(r#"{ value(values: ["a", "b", "cdef"])}"#)
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "[String!]": the value length is 3, must be less than or equal to 2"#.to_string(),
+            source: None,
+            locations: vec![Pos { column: 17, line: 1}],
+            path: vec![PathSegment::Field("value".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value(values: ["a", "cdef"])}"#)
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "String": the string length is 4, must be less than or equal to 3"#.to_string(),
+            source: None,
+            locations: vec![Pos { column: 17, line: 1}],
+            path: vec![PathSegment::Field("value".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value(values: ["a", "b"])}"#)
+            .await
+            .into_result()
+            .unwrap()
+            .data,
+        value!({
+            "value": "ab"
+        })
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value2(values: ["a", "b", "cdef"])}"#)
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "[String!]": the value length is 3, must be less than or equal to 2"#.to_string(),
+            source: None,
+            locations: vec![Pos { column: 18, line: 1}],
+            path: vec![PathSegment::Field("value2".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value2(values: ["a", "cdef"])}"#)
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "String": the string length is 4, must be less than or equal to 3"#.to_string(),
+            source: None,
+            locations: vec![Pos { column: 18, line: 1}],
+            path: vec![PathSegment::Field("value2".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value2(values: ["a", "b", "cdef"])}"#)
+            .await
+            .into_result()
+            .unwrap_err(),
+        vec![ServerError {
+            message: r#"Failed to parse "[String!]": the value length is 3, must be less than or equal to 2"#.to_string(),
+            source: None,
+            locations: vec![Pos { column: 18, line: 1}],
+            path: vec![PathSegment::Field("value2".to_string())],
+            extensions: None
+        }]
+    );
+
+    assert_eq!(
+        schema
+            .execute(r#"{ value2(values: null)}"#)
+            .await
+            .into_result()
+            .unwrap()
+            .data,
+        value!({
+            "value2": ""
+        })
     );
 }